+++ src/main/Application.js
@@ -4,9 +4,9 @@ const React = require('react')
     , h = require('react-hyperscript')
     , through = require('through2')
     , Immutable = require('immutable')
-    , NotFound = require('./not_found')
-    , Footer = require('./footer')
-    , Header = require('./header')
+    , NotFound = require('./NotFound')
+    , Footer = require('./Footer')
+    , Header = require('./Header')
     , { connect } = require('react-redux')
     , locationHashStream = require('location-hash-stream')
     , { match } = require('../router')
